requirements.txt
pandas
seaborn
requests
plotly
ipython
nbformat
kaleido
psutil
azure-storage-file-datalake --pre
azure-identity
pydrive